remove usRouteData hook for uneeded rerenders
