+++ client/src/components/Form.tsx
@@ -1,11 +1,14 @@
 import { Link } from "react-router-dom";
-import useRouteHandler from "../hooks/useRouteHandler";
+import getRouteData from "../utilities/getRouteData";
+import { useLocation } from "react-router-dom";
 import useAPI from "../hooks/useAPI";
 import "./Form.css";
 
 const Form: React.FC = (): React.ReactElement => {
+  console.log("Form");
+  const location = useLocation();
   const [formData, setFormData, APICall] = useAPI();
-  const routeData = useRouteHandler();
+  const routeData = getRouteData(location.pathname);
 
   const inputHandler = (e: React.ChangeEvent<HTMLInputElement>): void => {
     const { name, value } = e.target;
