+++ client/src/main.tsx
@@ -1,16 +1,13 @@
-import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App.jsx";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
-  <StrictMode>
-    <BrowserRouter>
-      <div id="background-image">
-        <div id="background-radial" />
-      </div>
-      <App />
-    </BrowserRouter>
-  </StrictMode>
+  <BrowserRouter>
+    <div id="background-image">
+      <div id="background-radial" />
+    </div>
+    <App />
+  </BrowserRouter>
 );
